Run test setup just once for integration tests
We were running the setup of the tests on each test for
each service. This was expensive since we need to load heavy data
structures from the System such as `PublicParameters` on each
test and it takes quite some time.

So as it was stated in #144 the integration test module has been
refactored so that the setup is just done once.

Closes #144
