+++ Cargo.toml
@@ -31,7 +31,7 @@ tracing = "0.1"
 tracing-subscriber = "0.2"
 clap = "2.33.3"
 prost = "0.6"
-futures = { version = "0.3", default-features = false, features = ["alloc"] }
+futures = "0.3"
 thiserror = "1.0"
 anyhow = "1.0"
 rustc_tools_util = "0.2"
