+++ tests/services/mod.rs
@@ -1,0 +1,3 @@
+pub mod blindbid_service;
+pub mod echo_service;
+pub mod pki_service;
